% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dheatmap.R
\name{dheatmap}
\alias{dheatmap}
\title{Heatmap Plot}
\usage{
dheatmap(d_matrix, colors = c("white", "lightblue", "blue", "darkblue",
  "black"), mode = "linear")
}
\arguments{
\item{d_matrix}{A matrix of D statistics, given by the output of dwrapper.}

\item{colors}{A color vector. Optionally modify the color scheme of the heatmap. If mode = 'binned', must be of length 5.}

\item{mode}{A string indicating desired mode. Use "linear" for most applications including use on base D statistic matrices.
For either of the ratio matrices, 'truncated' is most appropriate, though "binned" may be used if you would prefer a more
discreet distribution of colors.}
}
\value{
A color plot
}
\description{
Displays results of a D statistic scan as a heatmap.
}
