% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwrapper.R
\name{dwrapper}
\alias{dwrapper}
\title{Ohta D Statistic Wrapper}
\usage{
dwrapper(data_set, tot_maf = 0.1, pop_maf = 0.05)
}
\arguments{
\item{data_set}{Matrix containing genotype data with individuals or lines as rows and loci as columns}

\item{tot_maf}{Minimum frequency each polymorphism must have in the total population in order to be included in analysis}

\item{pop_maf}{Minimum frequency each polymorphism must have in a subpopulation in order for that subpopulation to be included in analysis}
}
\value{
A list of matrices containing the pairwise comparisons for each D statistic. Also included is the number of subpopulations evaluated in
each comparison and the ratio of d2is_mat to d2st_mat (ratio1) and dp2st_mat to dp2is_mat (ratio2).
}
\description{
Pairwise computation of Ohta's D Statistics for each polymorphism in a given dataset
}
