% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miyashita_langley_data.R
\docType{data}
\name{miyashita_langley_data}
\alias{miyashita_langley_data}
\title{Drosophila melanogaster genotypes}
\format{A matrix with 64 rows and 85 columns.}
\usage{
data(miyashita_langley_data)
}
\description{
Genotype data obtained from Miyashita & Langley (1988). A matrix representing 85 loci
in 64 individuals. Individuals are in rows. Rownames "RL", "TX", or "FK", indicate
the subpopulation from which the sample was taken.
}
\references{
Miyashita & Langley (1988) Genetics 120:199-212
(https://www.ncbi.nlm.nih.gov/pmc/articles/PMC1203490/)
}
\keyword{datasets}
