% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohtadstat.R
\name{ohtadstat}
\alias{ohtadstat}
\title{Tomoka Ohta's D Statistics}
\usage{
ohtadstat(index, data_set, tot_maf = 0.1, pop_maf = 0.05)
}
\arguments{
\item{index}{Vector of column names for which Ohta's D Statistics will be computed}

\item{data_set}{Matrix containing genotype data with individuals or lines as rows and loci as columns}

\item{tot_maf}{Minimum frequency a polymorphism must have in the total population in order to be included in analysis}

\item{pop_maf}{Minimum frequency a polymorphism must have in a subpopulation in order for that subpopulation to be included in analysis}
}
\value{
nPops Number of subpopulations present in the data set

D2it Measure of the correlation of the alleles at the specified loci

D2is Expected variance of LD in all subpopulations

Dp2st Variance of LD for the total population computed over alleles only

Dp2is Correlation of alleles at the specified loci relative to their expected correlation in the total population
}
\description{
Implements Ohta's D statistics for a pair of loci
}
