% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelprep.R
\name{parallelprep}
\alias{parallelprep}
\title{Make prep file for parallelization}
\usage{
parallelprep(data_set, outfile = "ohtacomparisons.txt")
}
\arguments{
\item{data_set}{The data set that is to be analysed.}

\item{outfile}{The name of the file to save the table to. May also be a path.}
}
\description{
Builds a table of pairwise locus comparisons that need to be made. This table can then be used by a workload manager such as 
slurm or condor to distribute the tasks among many instances of the dstat function.
}
\examples{
data(beissinger_data)
parallelprep(data_set = beissinger_data, outfile = "beissinger_comparisons.txt")

}
